backend/requirements_prod.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-yasg==1.21.7
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.0.1
channels==4.0.0
channels-redis==4.1.0
